Add GitHub Configuration Properties file
