+++ src/main/java/aws/lambda/micronaut/workshop/github/GitHubResponse.java
@@ -1,4 +1,4 @@
-package aws.lambda.micronaut.workshop;
+package aws.lambda.micronaut.workshop.github;
 
 import io.micronaut.core.annotation.Introspected;
 import lombok.Data;
